Refactor: Split USBIPMessages.swift to resolve SwiftLint file length violations
- Split 847-line USBIPMessages.swift into focused modules:
  - USBIPHeader.swift (42 lines): Common header structure
  - USBIPDeviceMessages.swift (273 lines): Device list and import messages
  - USBIPTransferMessages.swift (382 lines): USB transfer messages
  - USBIPMessages.swift (200 lines): Utilities and protocol definitions

- All files now comply with SwiftLint 500-line limit
- Maintains complete functional compatibility
- Improves code organization and maintainability
